requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
appnope==0.1.4
asttokens==3.0.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
comm==0.2.3
debugpy==1.8.17
decorator==5.2.1
executing==2.2.1
fastapi==0.104.1
Flask==2.3.3
google-api-core==2.25.2
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
graphviz==0.21
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
ipykernel==6.30.1
ipython==9.6.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jupyter_client==8.6.3
jupyter_core==5.8.1
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib-inline==0.1.7
narwhals==2.7.0
nest-asyncio==1.6.0
numpy==2.2.5
oauthlib==3.3.1
packaging==25.0
pandas==2.2.3
parso==0.8.5
pexpect==4.9.0
platformdirs==4.5.0
plotly==6.3.1
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==6.32.1
psutil==7.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.5.0
pydantic_core==2.14.1
Pygments==2.19.2
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2025.2
pyzmq==27.1.0
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9.1
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.27.0
tornado==6.5.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
wcwidth==0.2.14
Werkzeug==3.1.3
